requirements.txt
fastapi==0.112.1
uvicorn==0.30.6
sqlalchemy==2.0.32
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
fastapi-pagination==0.12.24
python-socketio==5.11.4
websockets==13.1
httpx==0.27.0
bcrypt==4.2.0
pytest==8.3.2
pytest-asyncio==0.23.5
fastapi-utils[all]==0.8.0
PyJWT==2.8.0
